refact : thread safe한 방향으로 변경 함수의 매개변수에 줘서 thread safe하도록 함 이때 커넥션 객체는 동일한걸 사용
