+++ src/main/java/boost/chaper2/DatabaseConnector.java
@@ -1,0 +1,27 @@
+package boost.chaper2;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DatabaseConnector {
+    private static String url = "jdbc:mysql://localhost:3306/tables_in_connectdb";
+    private static String user = "boost";
+    private static String password = "boost";
+
+
+    static {
+        try {
+            Class.forName("com.mysql.cj.jdbc.Driver");
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static Connection getConnection() throws SQLException {
+        return DriverManager.getConnection(url, user, password);
+    }
+
+
+
+}
